Another fix to a persistent bug in polytree nesting (#618)
